reorg: move some functions out of tagManipulation module; rename modules to match function
