+++ src/components/projects/CardContainer.jsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react";
 import { useStore } from "@nanostores/react";
 import { selectedTags } from "@stores/selectedTagsStore";
 import { selectedProjectType } from "@stores/selectedProjectTypeStore";
-import { extractUniqueTagValueArray } from "@utils/tagManipulation";
+import { extractUniqueTagValueArray } from "@utils/extractUniqueTags";
 import { $projectData } from "@stores/projectSearchResultsStore";
 import { $ecosystemData } from "@stores/ecosystemSearchResultsStore";
 import { $urlQuery } from "@stores/queryStore";
