+++ src/components/projects/Tag.astro
@@ -1,6 +1,6 @@
 ---
-import { getBackgroundColor } from "@utils/tagManipulation";
-import { updateSearchParamUrl } from "@utils/tagManipulation";
+import { getBackgroundColor } from "@utils/getTagBgColor";
+import { updateSearchParamUrl } from "@utils/updateTagUrlParams";
 
 const { tagCat, tag, contentType } = Astro.props;
