+++ src/pages/ecosystems/index.astro
@@ -14,7 +14,7 @@ import CardContent from "@components/projects/CardContent.astro";
 import SearchInput from "@components/projects/SearchInput";
 import Tag from "@components/projects/Tag.astro";
 
-import { extractUniqueTagsObject } from "@utils/tagManipulation.js";
+import { extractUniqueTagsObject } from "@utils/extractUniqueTags.js";
 
 //allEcosystems gets passes as "content" to the content cards
 //eveything else happening in here is to make tags for the filter menu and cards
