+++ src/utils/extractUniqueTags.js
@@ -1,31 +1,4 @@
 import validTagsList from "../../.github/actions/validTagsList.json";
-import possibleTagColors from "./tagColors.json";
-import { allLabNamesAndUrls } from "@data/labNamesUrls";
-
-// Used in the filter menu to add/remove params
-// Used when clicking on a tag to remove all params and add only the one clicked
-export function updateSearchParamUrl(urlStringToUpdate, tagInput) {
-  const url = new URL(urlStringToUpdate);
-  const searchParams = new URLSearchParams(url.search);
-
-  searchParams.delete("tag");
-
-  if (Array.isArray(tagInput)) {
-    tagInput.forEach((tag) => searchParams.append("tag", tag));
-  } else {
-    searchParams.append("tag", tagInput);
-  }
-
-  return `${url.origin}${url.pathname}?${searchParams.toString()}`;
-}
-
-// Used for assigning tag background colors in the filter menu
-export function getBackgroundColor(key) {
-  const index = Object.keys(validTagsList).indexOf(key);
-  // Use the index to return the color at the same index in possibleTagColors
-  // If the index is not found (-1), return a default color
-  return index >= 0 ? possibleTagColors[index] : "#6ebebd";
-}
 
 // Takes in a content collection object (e.g., projects or ecosystems) and returns
 // an object of only the unique tag keys and unique tag values within each key
@@ -35,12 +8,12 @@ export function getBackgroundColor(key) {
 export function extractUniqueTagsObject(contentCollection) {
   const uniqueTags = {};
   const validTagKeys = Object.keys(validTagsList);
-
   // For each key, determine the union of all unique values associated with it
   // (across all content items), excluding keys which aren't in validTagsList.
   contentCollection = Array.isArray(contentCollection)
     ? contentCollection
     : [contentCollection];
+
   contentCollection.forEach((contentItem) => {
     Object.entries(contentItem.data).forEach(([key, tagValues]) => {
       if (!validTagKeys.includes(key) || !tagValues || !tagValues.length) {
@@ -76,20 +49,7 @@ export function extractUniqueTagValueArray(projectData) {
       }
     }
   });
+
   const tagValuesArray = Object.values(uniqueTagsObj).flat();
   return tagValuesArray;
-}
-
-//Used to take the "associated labs and projects" tag and find the associated URL
-export function findLabInfo(labNames) {
-  if (!Array.isArray(labNames)) {
-    labNames = [labNames];
-  }
-  const labInfoArray = labNames.map((labName) => {
-    const labData = allLabNamesAndUrls.find((entry) =>
-      entry[0].includes(labName)
-    );
-    return labData ? { name: labName, url: labData[1] } : null;
-  });
-  return labInfoArray;
 }
