--- src/utils/tagColors.json
@@ -1,1 +1,0 @@
-["var(--color-primary)", "#ab3f32", "#8b8883", "#5084ac", "#ba5915", "#7b0249", "var(--color-accent)", "#24a3cc", "#b258b2"]
